LayoutTests/imported/w3c/web-platform-tests/html/rendering/non-replaced-elements/the-fieldset-and-legend-elements/legend-grid-flex-multicol-expected.txt

12

1
2

1
2

1
2

1
2

1
2

FAIL flex assert_equals: offsetHeight expected 18 but got 36
FAIL inline-flex assert_equals: offsetHeight expected 18 but got 36
FAIL grid assert_equals: offsetHeight expected 18 but got 36
FAIL inline-grid assert_equals: offsetHeight expected 18 but got 36
PASS multicol
